+++ crates/runtime/tests/object_tests.rs
@@ -50,13 +50,13 @@ mod objects {
     }
 
     macro_rules! arithmetic_op {
-        ($self:ident, $rhs:expr, $op:tt) => {
+        ($self:ident, $other:expr, $op:tt) => {
             {
                 use KValue::*;
-                match $rhs {
-                    Object(rhs) if rhs.is_a::<Self>() => {
-                        let rhs = rhs.cast::<Self>().unwrap();
-                        Ok(Self::make_value($self.x $op rhs.x))
+                match $other {
+                    Object(other) if other.is_a::<Self>() => {
+                        let other = other.cast::<Self>().unwrap();
+                        Ok(Self::make_value($self.x $op other.x))
                     }
                     Number(n) => {
                         Ok(Self::make_value($self.x $op i64::from(n)))
@@ -69,14 +69,29 @@ mod objects {
         }
     }
 
+    macro_rules! arithmetic_op_rhs {
+        ($self:ident, $other:expr, $op:tt) => {
+            {
+                match $other {
+                    KValue::Number(n) => {
+                        Ok(Self::make_value(i64::from(n) $op $self.x))
+                    }
+                    unexpected => {
+                        unexpected_type(&format!("a {} or Number", Self::type_static()), unexpected)
+                    }
+                }
+            }
+        }
+    }
+
     macro_rules! assignment_op {
-        ($self:ident, $rhs:expr, $op:tt) => {
+        ($self:ident, $other:expr, $op:tt) => {
             {
                 use KValue::*;
-                match $rhs {
-                    Object(rhs) if rhs.is_a::<Self>() => {
-                        let rhs = rhs.cast::<Self>().unwrap();
-                        $self.x $op rhs.x;
+                match $other {
+                    Object(other) if other.is_a::<Self>() => {
+                        let other = other.cast::<Self>().unwrap();
+                        $self.x $op other.x;
                         Ok(())
                     }
                     Number(n) => {
@@ -92,14 +107,14 @@ mod objects {
     }
 
     macro_rules! comparison_op {
-        ($self:ident, $rhs:expr, $op:tt) => {
+        ($self:ident, $other:expr, $op:tt) => {
             {
                 use KValue::*;
-                match $rhs {
-                    Object(rhs) if rhs.is_a::<Self>() => {
-                        let rhs = rhs.cast::<Self>().unwrap();
+                match $other {
+                    Object(other) if other.is_a::<Self>() => {
+                        let other = other.cast::<Self>().unwrap();
                         #[allow(clippy::float_cmp)]
-                        Ok($self.x $op rhs.x)
+                        Ok($self.x $op other.x)
                     }
                     Number(n) => {
                         #[allow(clippy::float_cmp)]
@@ -181,68 +196,88 @@ mod objects {
             Ok(Self::make_value(-self.x))
         }
 
-        fn add(&self, rhs: &KValue) -> Result<KValue> {
-            arithmetic_op!(self, rhs, +)
+        fn add(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op!(self, other, +)
         }
 
-        fn subtract(&self, rhs: &KValue) -> Result<KValue> {
-            arithmetic_op!(self, rhs, -)
+        fn add_rhs(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op_rhs!(self, other, +)
         }
 
-        fn multiply(&self, rhs: &KValue) -> Result<KValue> {
-            arithmetic_op!(self, rhs, *)
+        fn subtract(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op!(self, other, -)
         }
 
-        fn divide(&self, rhs: &KValue) -> Result<KValue> {
-            arithmetic_op!(self, rhs, /)
+        fn subtract_rhs(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op_rhs!(self, other, -)
         }
 
-        fn remainder(&self, rhs: &KValue) -> Result<KValue> {
-            arithmetic_op!(self, rhs, %)
+        fn multiply(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op!(self, other, *)
         }
 
-        fn add_assign(&mut self, rhs: &KValue) -> Result<()> {
-            assignment_op!(self, rhs, +=)
+        fn multiply_rhs(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op_rhs!(self, other, *)
         }
 
-        fn subtract_assign(&mut self, rhs: &KValue) -> Result<()> {
-            assignment_op!(self, rhs, -=)
+        fn divide(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op!(self, other, /)
         }
 
-        fn multiply_assign(&mut self, rhs: &KValue) -> Result<()> {
-            assignment_op!(self, rhs, *=)
+        fn divide_rhs(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op_rhs!(self, other, /)
         }
 
-        fn divide_assign(&mut self, rhs: &KValue) -> Result<()> {
-            assignment_op!(self, rhs, /=)
+        fn remainder(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op!(self, other, %)
         }
 
-        fn remainder_assign(&mut self, rhs: &KValue) -> Result<()> {
-            assignment_op!(self, rhs, %=)
+        fn remainder_rhs(&self, other: &KValue) -> Result<KValue> {
+            arithmetic_op_rhs!(self, other, %)
         }
 
-        fn less(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, <)
+        fn add_assign(&mut self, other: &KValue) -> Result<()> {
+            assignment_op!(self, other, +=)
         }
 
-        fn less_or_equal(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, <=)
+        fn subtract_assign(&mut self, other: &KValue) -> Result<()> {
+            assignment_op!(self, other, -=)
         }
 
-        fn greater(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, >)
+        fn multiply_assign(&mut self, other: &KValue) -> Result<()> {
+            assignment_op!(self, other, *=)
         }
 
-        fn greater_or_equal(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, >=)
+        fn divide_assign(&mut self, other: &KValue) -> Result<()> {
+            assignment_op!(self, other, /=)
         }
 
-        fn equal(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, ==)
+        fn remainder_assign(&mut self, other: &KValue) -> Result<()> {
+            assignment_op!(self, other, %=)
         }
 
-        fn not_equal(&self, rhs: &KValue) -> Result<bool> {
-            comparison_op!(self, rhs, !=)
+        fn less(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, <)
+        }
+
+        fn less_or_equal(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, <=)
+        }
+
+        fn greater(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, >)
+        }
+
+        fn greater_or_equal(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, >=)
+        }
+
+        fn equal(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, ==)
+        }
+
+        fn not_equal(&self, other: &KValue) -> Result<bool> {
+            comparison_op!(self, other, !=)
         }
 
         fn is_iterable(&self) -> IsIterable {
@@ -503,8 +538,9 @@ make_object(10)
         #[test]
         fn add() {
             let script = "
-x = (make_object 11) + (make_object 22) + 33
-x.as_number()
+x = (make_object 11) + (make_object 22)
+y = 33 + x
+y.as_number()
 ";
             test_object_script(script, 66);
         }
@@ -512,8 +548,9 @@ x.as_number()
         #[test]
         fn subtract() {
             let script = "
-x = (make_object 99) - (make_object 90) - 9
-x.as_number()
+x = (make_object 99) - (make_object 90) - 1
+y = 8 - x
+y.as_number()
 ";
             test_object_script(script, 0);
         }
@@ -522,27 +559,30 @@ x.as_number()
         fn multiply() {
             let script = "
 x = (make_object 3) * (make_object 11)
-x.as_number()
+y = 10 * x
+y.as_number()
 ";
-            test_object_script(script, 33);
+            test_object_script(script, 330);
         }
 
         #[test]
         fn divide() {
             let script = "
 x = (make_object 90) / (make_object 10)
-x.as_number()
+y = 9 / x
+y.as_number()
 ";
-            test_object_script(script, 9);
+            test_object_script(script, 1);
         }
 
         #[test]
         fn remainder() {
             let script = "
 x = (make_object 45) % (make_object 10)
-x.as_number()
+y = 12 % x
+y.as_number()
 ";
-            test_object_script(script, 5);
+            test_object_script(script, 2);
         }
 
         #[test]