+++ libs/geometry/src/macros.rs
@@ -61,12 +61,12 @@ macro_rules! impl_arithmetic_ops {
 
 #[macro_export]
 macro_rules! geometry_arithmetic_op {
-    ($self:ident, $rhs:expr, $op:tt) => {
+    ($self:ident, $other:expr, $op:tt) => {
         {
-            match $rhs {
-                KValue::Object(rhs) if rhs.is_a::<Self>() => {
-                    let rhs = rhs.cast::<Self>().unwrap();
-                    Ok((*$self $op *rhs).into())
+            match $other {
+                KValue::Object(other) if other.is_a::<Self>() => {
+                    let other = other.cast::<Self>().unwrap();
+                    Ok((*$self $op *other).into())
                 }
                 KValue::Number(n) => {
                     Ok((*$self $op f64::from(n)).into())
@@ -80,13 +80,29 @@ macro_rules! geometry_arithmetic_op {
 }
 
 #[macro_export]
-macro_rules! geometry_compound_assign_op {
-    ($self:ident, $rhs:expr, $op:tt) => {
+macro_rules! geometry_arithmetic_op_rhs {
+    ($self:ident, $other:expr, $op:tt) => {
         {
-            match $rhs {
-                KValue::Object(rhs) if rhs.is_a::<Self>() => {
-                    let rhs = rhs.cast::<Self>().unwrap();
-                    *$self $op *rhs;
+            match $other {
+                KValue::Number(n) => {
+                    Ok((Self::from(f64::from(n)) $op *$self).into())
+                }
+                unexpected => {
+                    unexpected_type(&format!("a {} or Number", Self::type_static()), unexpected)
+                }
+            }
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! geometry_compound_assign_op {
+    ($self:ident, $other:expr, $op:tt) => {
+        {
+            match $other {
+                KValue::Object(other) if other.is_a::<Self>() => {
+                    let other = other.cast::<Self>().unwrap();
+                    *$self $op *other;
                     Ok(())
                 }
                 KValue::Number(n) => {
@@ -103,12 +119,12 @@ macro_rules! geometry_compound_assign_op {
 
 #[macro_export]
 macro_rules! geometry_comparison_op {
-    ($self:ident, $rhs:expr, $op:tt) => {
+    ($self:ident, $other:expr, $op:tt) => {
         {
-            match $rhs {
-                KValue::Object(rhs) if rhs.is_a::<Self>() => {
-                    let rhs = rhs.cast::<Self>().unwrap();
-                    Ok(*$self $op *rhs)
+            match $other {
+                KValue::Object(other) if other.is_a::<Self>() => {
+                    let other = other.cast::<Self>().unwrap();
+                    Ok(*$self $op *other)
                 }
                 unexpected => {
                     unexpected_type(&format!("a {}", Self::type_static()), unexpected)