+++ libs/geometry/src/vec3.rs
@@ -43,44 +43,56 @@ impl KotoObject for Vec3 {
         Ok(Self(-self.0).into())
     }
 
-    fn add(&self, rhs: &KValue) -> Result<KValue> {
-        geometry_arithmetic_op!(self, rhs, +)
+    fn add(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op!(self, other, +)
     }
 
-    fn subtract(&self, rhs: &KValue) -> Result<KValue> {
-        geometry_arithmetic_op!(self, rhs, -)
+    fn add_rhs(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op_rhs!(self, other, +)
     }
 
-    fn multiply(&self, rhs: &KValue) -> Result<KValue> {
-        geometry_arithmetic_op!(self, rhs, *)
+    fn subtract(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op!(self, other, -)
     }
 
-    fn divide(&self, rhs: &KValue) -> Result<KValue> {
-        geometry_arithmetic_op!(self, rhs, /)
+    fn subtract_rhs(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op_rhs!(self, other, -)
     }
 
-    fn add_assign(&mut self, rhs: &KValue) -> Result<()> {
-        geometry_compound_assign_op!(self, rhs, +=)
+    fn multiply(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op!(self, other, *)
     }
 
-    fn subtract_assign(&mut self, rhs: &KValue) -> Result<()> {
-        geometry_compound_assign_op!(self, rhs, -=)
+    fn multiply_rhs(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op_rhs!(self, other, *)
     }
 
-    fn multiply_assign(&mut self, rhs: &KValue) -> Result<()> {
-        geometry_compound_assign_op!(self, rhs, *=)
+    fn divide(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op!(self, other, /)
     }
 
-    fn divide_assign(&mut self, rhs: &KValue) -> Result<()> {
-        geometry_compound_assign_op!(self, rhs, /=)
+    fn divide_rhs(&self, other: &KValue) -> Result<KValue> {
+        geometry_arithmetic_op_rhs!(self, other, /)
     }
 
-    fn equal(&self, rhs: &KValue) -> Result<bool> {
-        geometry_comparison_op!(self, rhs, ==)
+    fn add_assign(&mut self, other: &KValue) -> Result<()> {
+        geometry_compound_assign_op!(self, other, +=)
     }
 
-    fn not_equal(&self, rhs: &KValue) -> Result<bool> {
-        geometry_comparison_op!(self, rhs, !=)
+    fn subtract_assign(&mut self, other: &KValue) -> Result<()> {
+        geometry_compound_assign_op!(self, other, -=)
+    }
+
+    fn multiply_assign(&mut self, other: &KValue) -> Result<()> {
+        geometry_compound_assign_op!(self, other, *=)
+    }
+
+    fn divide_assign(&mut self, other: &KValue) -> Result<()> {
+        geometry_compound_assign_op!(self, other, /=)
+    }
+
+    fn equal(&self, other: &KValue) -> Result<bool> {
+        geometry_comparison_op!(self, other, ==)
     }
 
     fn index(&self, index: &KValue) -> Result<KValue> {
@@ -119,6 +131,12 @@ impl KotoObject for Vec3 {
 impl From<DVec3> for Vec3 {
     fn from(v: DVec3) -> Self {
         Self(v)
+    }
+}
+
+impl From<f64> for Vec3 {
+    fn from(x: f64) -> Self {
+        Self::new(x, x, x)
     }
 }